add put("/backends/{collection}") endpoint
This endpoint creates a backend document in the 'collection' name
passed in the request. If the document containing the backend
name already present then the document is updated. any key value
pairs can be updated by a dictionary containing  the 'name' filed
and required update of keys:values. Any change i.e. creation or
updating a backend document is logged in hardcoded 'backend_log'
collection.